Part4

+++ osa-4/blogilista/controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blogi')
 const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
@@ -8,49 +10,75 @@ blogRouter.get('/', async (request, response) => {
   response.json(blogs.map((blog) => blog.toJSON()))
 })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', userExtractor, async (request, response) => {
   const { body } = request
+  const { token } = request
+  const { user } = request
 
   if (body.title === undefined || body.url === undefined) {
     return response.status(400).json({ error: "Title or/and url can't be empty" }).end()
   }
 
-  const users = await User.find({})
+  if (!token || user === undefined) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
-    user: users[0]._id,
+    user: user._id,
   })
 
   const blogData = await blog.save()
-  users[0].blogs = users[0].blogs.concat(blogData._id)
-  await users[0].save()
+  user.blogs = user.blogs.concat(blogData._id)
+  await user.save()
 
   return response.status(201).json(blogData)
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+  const { token } = request
+  const { user } = request
+
+  if (!token || user === undefined) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'Only the creator of this blog can delete it' })
+  }
 })
 
-blogRouter.put('/:id', async (request, response) => {
+blogRouter.put('/:id', userExtractor, async (request, response) => {
+  const { token } = request
   const { body } = request
+  const { user } = request
+
+  if (!token || user === undefined) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
 
   try {
     const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === user.id.toString()) {
+      const newBlog = {
+        title: body.title || blog.title,
+        author: body.author || blog.author,
+        url: body.url || blog.url,
+        likes: body.likes || blog.likes,
+      }
+      const updateBlog = await Blog.findByIdAndUpdate(request.params.id, newBlog, { new: true })
 
-    const newBlog = {
-      title: body.title || blog.title,
-      author: body.author || blog.author,
-      url: body.url || blog.url,
-      likes: body.likes || blog.likes,
+      response.status(202).json(updateBlog.toJSON())
+    } else {
+      return response.status(401).json({ error: 'Only the creator of this blog can edit this blog' })
     }
-    const updateBlog = await Blog.findByIdAndUpdate(request.params.id, newBlog, { new: true })
-
-    response.status(202).json(updateBlog.toJSON())
   } catch (error) {
     response.status(400).json(error).end()
   }
